pulling out some functions that will be used by reprocessing code into a separate python package.  Moving around unit tests.
